+++ src/main/java/com/scm/app/model/TimeTable.java
@@ -40,13 +40,13 @@ public class TimeTable {
 	@OneToMany(targetEntity = DayTimeTable.class, cascade = CascadeType.ALL)
 	List<DayTimeTable> dayTimeTable;
 
-	private Integer accountId;
+	private Long accountId;
 
-	public Integer getAccountId() {
+	public Long getAccountId() {
 		return accountId;
 	}
 
-	public void setAccountId(Integer accountId) {
+	public void setAccountId(Long accountId) {
 		this.accountId = accountId;
 	}
