+++ src/main/java/com/scm/app/repo/StudentRepo.java
@@ -12,5 +12,5 @@ public interface StudentRepo extends JpaRepository<Student, Long> {
 
 	List<Student> getByDivisionIdAndClassId(Long divisionId, Long classId);
 
-    long countByAccountId(Integer accountId);
+    long countByAccountId(Long accountId);
 }
